+++ SE_DB/dataaccess/src/main/java/by/grodno/zagart/dataaccess/dao/mapper/OrderMapper.java
@@ -19,8 +19,8 @@ public class OrderMapper implements RowMapper<Order> {
 	@Override
 	public Order mapRow(ResultSet rs, int rowNum) throws SQLException {
 		Order order = new Order();
-		long id = rs.getLong("id");
-		long shoppingCartId = rs.getLong("shopping_cart_id");
+		int id = rs.getInt("id");
+		Integer shoppingCartId = rs.getInt("shopping_cart_id");
 		Date dateOfOrder = rs.getDate("date_of_order");
 		Date dateOfDeliver = rs.getDate("date_of_deliver");
 		String orderStatus = rs.getString("order_status");
